Updated test to JUnit5 to remove reference to Dialect subclass Moved it back for now to the old location, since it fails with a non-related error that will have to be looked at as soon as there is an implementation for composite sub-types
